Refactor social links management: update appData and homePageData structures, implement async loading in App component, and remove redundant social links from HomePage.
