+++ public/assets/data/appData.json
@@ -1,3 +1,20 @@
 {
-  "appHeaderText": "Portfolio"
+  "appHeaderText": "Portfolio",
+  "socialLinks": [
+    {
+      "name": "GitHub",
+      "url": "https://github.com/javidahmed64592",
+      "icon": "github.svg"
+    },
+    {
+      "name": "LinkedIn",
+      "url": "https://www.linkedin.com/in/javidahmed64592",
+      "icon": "linkedin.svg"
+    },
+    {
+      "name": "Source Code",
+      "url": "https://github.com/javidahmed64592/portfolio",
+      "icon": "source.svg"
+    }
+  ]
 }
